move grocery lsit builder to end docs errors
